mood-food-app/backend/requirements.txt
flask[async]==3.0.2
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.1
gunicorn==21.2.0
transformers==4.36.2
torch==2.1.2
requests==2.31.0
asgiref==3.7.2 